Generate audit log entries for each request
