+++ plugin/plugin.go
@@ -8,9 +8,12 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"os"
+	"time"
 
 	"github.com/drone/drone-go/plugin/validator"
 	"github.com/open-policy-agent/opa/rego"
+	"github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
 )
 
@@ -21,28 +24,28 @@ func New(policy string) validator.Plugin {
 	}
 }
 
-type plugin struct {
-	policyPath string
-}
-
 func (p *plugin) Validate(ctx context.Context, req *validator.Request) error {
-	// Drone sends its raw config in the request body
-	type DroneConfig struct {
-		Kind  string `yaml:"kind" json:"kind"`
-		Type  string `yaml:"type" json:"type"`
-		Name  string `yaml:"name" json:"name"`
-		Steps []struct {
-			Name      string   `yaml:"name" json:"name"`
-			Image     string   `yaml:"image" json:"image"`
-			Commands  []string `yaml:"commands" json:"commands"`
-			DependsOn []string `yaml:"depends_on,omitempty" json:"depends_on,omitempty"`
-		} `yaml:"steps" json:"steps"`
-		Trigger struct {
-			Branch []string `yaml:"branch" json:"branch"`
-		} `yaml:"trigger" json:"trigger"`
-	}
 
 	var droneConfig DroneConfig
+
+	logrus.SetFormatter(&logrus.JSONFormatter{})
+	logrus.SetOutput(os.Stdout)
+
+	promotion := Promotion{
+		Build:       req.Build.Parent,
+		Environment: req.Build.Deploy,
+	}
+
+	// log each request received for audit purposes
+	logrus.WithFields(logrus.Fields{
+		"extension":  "validation",
+		"user":       req.Build.Sender,
+		"event":      req.Build.Event,
+		"repository": req.Repo.Slug,
+		"branch":     req.Repo.Branch,
+		"commit":     req.Build.After,
+		"promotion":  &promotion,
+	}).Info(fmt.Sprintf("audit(%d)", time.Now().Unix()))
 
 	err := yaml.Unmarshal([]byte(req.Config.Data), &droneConfig)
 	if err != nil {
@@ -65,7 +68,7 @@ func (p *plugin) Validate(ctx context.Context, req *validator.Request) error {
 
 	if rs[0].Bindings["deny"] == true {
 		message := fmt.Sprintf("[validator] %v", rs[0].Bindings["msg"])
-	 	return errors.New(message)
+		return errors.New(message)
 	}
 
 	if err != nil {
